# apps/backend/internal/domain/auth.go
package domain

import "time"

// RefreshToken represents a persisted refresh token session.
type RefreshToken struct {
	ID        string
	UserID    string
	TokenHash string
	ExpiresAt time.Time
	RevokedAt *time.Time
	CreatedAt time.Time
}

// TokenPair contains access and refresh tokens returned after authentication.
type TokenPair struct {
	AccessToken           string
	AccessTokenExpiresAt  time.Time
	RefreshToken          string
	RefreshTokenExpiresAt time.Time
}